Reset sitrep deck in the proper spot
At one point, negative sitreps were selected before positive ones on assault missions. The deck was reset after the first draw, so it was clean for the second. However, when the order was switched, the deck cleaning was not relocated and as a result we were drawing a lot more cards than were necessary.